fix: fix typo in example usage code (#47)
Example usage references `create_pandas_dataframe_agent` even though
this is the implementation for `create_csv_agent` and `create_csv_agent`
is what is imported in the example code on line 38.

As far as I understand, this is basically a wrapper around the
pandas_dataframe_agent, but still I think this is a typo in the example
code.